+++ components/src/core/header/headerActionItems.tsx
@@ -4,7 +4,7 @@ import Icon from 'react-native-vector-icons/MaterialIcons';
 import { wrapIcon } from '../icon-wrapper/icon-wrapper';
 import { HeaderIcon } from './headerIcon';
 import { useSearch } from './contexts/SearchContextProvider';
-import { HeaderAvatar, HeaderIcon as HeaderIconType } from '../__types__';
+import { HeaderActionComponent, HeaderIcon as HeaderIconType } from '../__types__';
 
 const ClearIcon = wrapIcon({ IconClass: Icon, name: 'clear' });
 const SearchIcon = wrapIcon({ IconClass: Icon, name: 'search' });
@@ -12,7 +12,7 @@ const SearchIcon = wrapIcon({ IconClass: Icon, name: 'search' });
 const makeStyles = (): StyleSheet.NamedStyles<{
     root: ViewStyle;
     actionItem: ViewStyle;
-    avatar: ViewStyle;
+    component: ViewStyle;
 }> => {
     const fontScale = PixelRatio.getFontScale();
 
@@ -30,21 +30,24 @@ const makeStyles = (): StyleSheet.NamedStyles<{
             paddingVertical: 8 * fontScale,
             paddingHorizontal: 8,
         },
-        avatar: {
+        component: {
             height: 40 * fontScale,
             width: 40 * fontScale,
+            justifyContent: 'center',
         },
     });
 };
 
 type ActionItemProps = {
     /** Array of up to three action items on the right of the header */
-    actionItems?: Array<HeaderIconType | HeaderAvatar>;
+    actionItems?: Array<HeaderIconType | HeaderActionComponent>;
 
     /** Style overrides for internal elements. The styles you provide will be combined with the default styles. */
     styles?: {
         root?: StyleProp<ViewStyle>;
         actionItem?: StyleProp<ViewStyle>;
+        component?: StyleProp<ViewStyle>;
+        /** TODO: Deprecated - remove this in 6.0.0, Add a changelog entry */
         avatar?: StyleProp<ViewStyle>;
     };
 };
@@ -53,15 +56,14 @@ type ActionItemProps = {
  * HeaderActionItems component
  *
  * The HeaderActionItems is a helper component for organizing the contents in the Header. It is
- * used for displaying all of the action item icons and avatars.
+ * used for displaying all of the action item icons and components.
  */
 export const HeaderActionItems: React.FC<ActionItemProps> = (props) => {
     const { actionItems, styles = {} } = props;
     const { searchConfig, searching, query, onClear, onSearch } = useSearch();
-    const MAX_ITEMS = 3;
     const defaultStyles = makeStyles();
 
-    let items: Array<HeaderIconType | HeaderAvatar> = actionItems || [];
+    let items: Array<HeaderIconType | HeaderActionComponent> = actionItems || [];
 
     if (searching) {
         if (query) {
@@ -89,13 +91,19 @@ export const HeaderActionItems: React.FC<ActionItemProps> = (props) => {
     if (items) {
         return (
             <View style={[defaultStyles.root, styles.root]}>
-                {items.slice(0, MAX_ITEMS).map((actionItem: HeaderIconType | HeaderAvatar, index) => {
+                {items.map((actionItem: HeaderIconType | HeaderActionComponent, index) => {
                     if ('component' in actionItem) {
                         return (
                             <View
                                 key={`action_${index}`}
                                 testID={`header-action-item${index}`}
-                                style={[defaultStyles.avatar, styles.avatar]}
+                                // TODO: Remove the avatar style in 6.0.0 - it's been deprecated
+                                style={[
+                                    defaultStyles.component,
+                                    actionItem.width ? { width: actionItem.width } : {},
+                                    styles.avatar,
+                                    styles.component,
+                                ]}
                             >
                                 {actionItem.component}
                             </View>